Move admin contracts into service core
