+++ Transcendence.Service.Core/Services/Admin/Implementations/AdminOverviewFacade.cs
@@ -8,9 +8,9 @@ using Microsoft.Extensions.Options;
 using Transcendence.Data;
 using Transcendence.Data.Models.LoL.Match;
 using Transcendence.Service.Core.Services.Admin.Interfaces;
+using Transcendence.Service.Core.Services.Admin.Models;
 using Transcendence.Service.Core.Services.Jobs.Configuration;
 using Transcendence.Service.Core.Services.RiotApi;
-using Transcendence.WebAPI.Controllers;
 
 namespace Transcendence.Service.Core.Services.Admin.Implementations;
