+++ roadmap.md
@@ -514,7 +514,7 @@ _Personalization, secondary surfaces, performance, and refactors_
   - **Fix:** Extract the acquire→enqueue→compensate flow into a single Service.Core orchestrator (e.g. ISummonerRefreshOrchestrator.QueueRefreshAsync) returning a small result the controllers map to IActionResult; both controllers call it.
   - **Why:** The lock TTL, priority-key semantics, compensation-on-enqueue-failure, and contention-response shape must be kept in sync by hand across two files; a fix or change to the lock protocol in one controller will silently diverge from the other, risking inconsistent refresh behavior (e.g. leaked locks) for one caller path.
   - **Where:** `Transcendence.WebAPI/Controllers/SummonersController.cs:304-389; Transcendence.WebAPI/Controllers/ProSummonersController.cs:222-299`
-- [ ] **Admin DTOs physically in Service.Core are declared under `namespace Transcendence.WebAPI.Controllers`, and facades `using Transcendence.WebAPI.Controllers`** `MED` · `medium`
+- [x] **Admin DTOs physically in Service.Core are declared under `namespace Transcendence.WebAPI.Controllers`, and facades `using Transcendence.WebAPI.Controllers`** `MED` · `medium`
   - **Fix:** Move the DTOs to a Service.Core-owned namespace (e.g. Transcendence.Service.Core.Services.Admin.Models) and, if the generated OpenAPI schema id must stay stable, pin it via Swashbuckle's SchemaId/CustomSchemaIds rather than by mislabeling the namespace.
   - **Why:** Namespace no longer reflects assembly/layer ownership. A reader (or static-analysis/dependency-direction tooling) sees Service.Core types and usings in a WebAPI namespace and cannot tell that the real dependency direction is intact. It also couples the Service.Core contract names to a WebAPI-shaped naming decision (OpenAPI schema stability), making a future namespace correction a breaking OpenAPI/client-regen change.
   - **Where:** `Transcendence.Service.Core/Services/Admin/Models/AdminOperationsContracts.cs:3-8; Transcendence.Service.Core/Services/Admin/Implementations/AdminJobsFacade.cs:9; Transcendence.Service.Core/Services/Admin/Implementations/AdminOverviewFacade.cs:13`
