+++ tests/Transcendence.WebAPI.Tests/AdminOperationsControllerTests.cs
@@ -14,6 +14,7 @@ using Transcendence.Data;
 using Transcendence.Data.Models.LoL.Match;
 using Transcendence.Data.Models.LoL.Static;
 using Transcendence.Service.Core.Services.Admin.Implementations;
+using Transcendence.Service.Core.Services.Admin.Models;
 using Transcendence.Service.Core.Services.Analytics.Interfaces;
 using Transcendence.Service.Core.Services.Auth.Interfaces;
 using Transcendence.Service.Core.Services.Diagnostics;
